Correct tyope of float

+++ lstchain/tools/lstchain_create_calibration_file.py
@@ -152,7 +152,7 @@ class CalibrationHDF5Writer(Tool):
                 # if simulation use not calibrated and not gain selected R0 waveform
                 if self.simulation:
                     event.r1.tel[tel_id].waveform = (
-                            event.r0.tel[tel_id].waveform.astype(float)
+                            event.r0.tel[tel_id].waveform.astype(np.float32)
                             -np.median(event.mon.tel[tel_id].calibration.pedestal_per_sample, axis=1)[...,np.newaxis, np.newaxis]
                     )
